condicion carrera

+++ src/main/java/edu/esi/ds/esientradas/dto/DtoEntradaInfo.java
@@ -1,6 +1,4 @@
 package edu.esi.ds.esientradas.dto;
 
-import java.math.BigDecimal;
-
-public record DtoEntradaInfo(BigDecimal total, BigDecimal libres, BigDecimal reservadas, BigDecimal vendidas) {
+public record DtoEntradaInfo(Long total, Long libres, Long reservadas, Long vendidas) {
 }
